[tools] tools ui popup: Allow returning hotkeys
Message-id: <20200522095131.24335-29-mpetlan@redhat.com>
Patchwork-id: 313656
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 28/98] tools ui popup: Allow returning hotkeys
Bugzilla: 1818841
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818841
https://bugzilla.redhat.com/show_bug.cgi?id=1818841

upstream
========
commit d07126560cab572539621702137eeeeb2a4edf30
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Dec 16 12:23:34 2019 -0300

description
===========
With this patch if an optional pointer is passed to ui__popup_menu()
then when any key that is not being handled (ENTER, ESC, etc) is typed,
it'll record that key in the pointer and return, allowing for hotkey
processing on the caller.

If NULL is passed, no change in logic, unhandled keys continue to be
ignored.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>