[tools] perf report/top: Make 'e' visible in the help and make it toggle showing callchains
Message-id: <20200522095131.24335-32-mpetlan@redhat.com>
Patchwork-id: 313658
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 31/98] perf report/top: Make 'e' visible in the help and make it toggle showing callchains
Bugzilla: 1818841
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818841
https://bugzilla.redhat.com/show_bug.cgi?id=1818841

upstream
========
commit e6d6abfc447a65e949c1e883e66e1450903d2fbd
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Dec 17 10:39:04 2019 -0300

description
===========
The 'e' and 'c' hotkeys were present for a long time, but not documented
in the help window, change 'e' to be a toggle so that it gets consistent
with other toggles like '+' and document it in the help window.

Keep 'c' as is for people used to it but don't document, as it is easier
to just use 'e' to show/hide all the callchains for a top level
histogram entry.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>