[tools] perf env: Do not return pointers to local variables
Message-id: <20200522095131.24335-81-mpetlan@redhat.com>
Patchwork-id: 313699
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 80/98] perf env: Do not return pointers to local variables
Bugzilla: 1818841
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1818841
https://bugzilla.redhat.com/show_bug.cgi?id=1818841

upstream
========
commit ebcb9464a2ae3a547e97de476575c82ece0e93e2
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Mar 2 11:23:03 2020 -0300

description
===========
It is possible to return a pointer to a local variable when looking up
the architecture name for the running system and no normalization is
done on that value, i.e. we may end up returning the uts.machine local
variable.

While this doesn't happen on most arches, as normalization takes place,
lets fix this by making that a static variable and optimize it a bit by
not always running uname(), only the first time.

Noticed in fedora rawhide running with:

  [perfbuilder@a5ff49d6e6e4 ~]$ gcc --version
  gcc (GCC) 10.0.1 20200216 (Red Hat 10.0.1-0.8)

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>