[powerpc] powerpc/pseries: Fix drc-info mappings of logical cpus to drc-index
Message-id: <20200522202026.10433-3-didoming@redhat.com>
Patchwork-id: 313787
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH BZ1779198 2/11] powerpc/pseries: Fix drc-info mappings of logical cpus to drc-index
Bugzilla: 1779198
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1779198
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28486934

There are a couple subtle errors in the mapping between cpu-ids and a
cpus associated drc-index when using the new ibm,drc-info property.

The first is that while drc-info may have been a supported firmware
feature at boot it is possible we have migrated to a CEC with older
firmware that doesn't support the ibm,drc-info property. In that case
the device tree would have been updated after migration to remove the
ibm,drc-info property and replace it with the older style ibm,drc-*
properties for types, indexes, names, and power-domains. PAPR even
goes as far as dictating that if we advertise support for drc-info
that we are capable of supporting either property type at runtime.

The second is that the first value of the ibm,drc-info property is
the int encoded count of drc-info entries. As such "value" returned
by of_prop_next_u32() is pointing at that count, and not the first
element of the first drc-info entry as is expected by the
of_read_drc_info_cell() helper.

Fix the first by ignoring DRC-INFO firmware feature and instead
testing directly for ibm,drc-info, and then falling back to the
old style ibm,drc-indexes in the case it doesn't exit.

Fix the second by incrementing value to the next element prior to
parsing drc-info entries.

commit 775fa495af04e0bdb3a00085aaa2d915ed51388f

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/1573449697-5448-3-git-send-email-tyreld@linux.ibm.com
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>