[pci] PCI: rpaphp: Don't rely on firmware feature to imply drc-info support
Message-id: <20200522202026.10433-6-didoming@redhat.com>
Patchwork-id: 313791
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH BZ1779198 5/11] PCI: rpaphp: Don't rely on firmware feature to imply drc-info support
Bugzilla: 1779198
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1779198
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28486934

In the event that the partition is migrated to a platform with older
firmware that doesn't support the ibm,drc-info property the device
tree is modified to remove the ibm,drc-info property and replace it
with the older style ibm,drc-* properties for types, names, indexes,
and power-domains. One of the requirements of the drc-info firmware
feature is that the client is able to handle both the new property,
and old style properties at runtime. Therefore we can't rely on the
firmware feature alone to dictate which property is currently
present in the device tree.

Fix this short coming by checking explicitly for the ibm,drc-info
property, and falling back to the older ibm,drc-* properties if it
doesn't exist.

commit 52e2b0f16574afd082cff0f0e8567b2d9f68c033

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>
Acked-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/1573449697-5448-6-git-send-email-tyreld@linux.ibm.com
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>