[pci] PCI: rpaphp: Correctly match ibm, my-drc-index to drc-name when using drc-info
Message-id: <20200522202026.10433-10-didoming@redhat.com>
Patchwork-id: 313796
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH BZ1779198 9/11] PCI: rpaphp: Correctly match ibm, my-drc-index to drc-name when using drc-info
Bugzilla: 1779198
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1779198
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28486934

The newer ibm,drc-info property is a condensed description of the old
ibm,drc-* properties (ie. names, types, indexes, and power-domains).
When matching a drc-index to a drc-name we need to verify that the
index is within the start and last drc-index range and map it to a
drc-name using the drc-name-prefix and logical index.

Fix the mapping by checking that the index is within the range of the
current drc-info entry, and build the name from the drc-name-prefix
concatenated with the starting drc-name-suffix value and the sequential
index obtained by subtracting ibm,my-drc-index from this entries
drc-start-index.

commit 4f9f2d3d7a434b7f882b72550194c9278f4a3925

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>
Acked-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/1573449697-5448-10-git-send-email-tyreld@linux.ibm.com
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>