[nvdimm] libnvdimm/region: Introduce an 'align' attribute
Message-id: <20200522210152.11736-6-didoming@redhat.com>
Patchwork-id: 313805
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH BZ1779218 5/5] libnvdimm/region: Introduce an 'align' attribute
Bugzilla: 1779218
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1779218
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28739413

The align attribute applies an alignment constraint for namespace
creation in a region. Whereas the 'align' attribute of a namespace
applied alignment padding via an info block, the 'align' attribute
applies alignment constraints to the free space allocation.

The default for 'align' is the maximum known memremap_compat_align()
across all archs (16MiB from PowerPC at time of writing) multiplied by
the number of interleave ways if there is blk-aliasing. The minimum is
PAGE_SIZE and allows for the creation of cross-arch incompatible
namespaces, just as previous kernels allowed, but the expectation is
cross-arch and mode-independent compatibility by default.

The regression risk with this change is limited to cases that were
dependent on the ability to create unaligned namespaces, *and* for some
reason are unable to opt-out of aligned namespaces by writing to
'regionX/align'. If such a scenario arises the default can be flipped
from opt-out to opt-in of compat-aligned namespace creation, but that is
a last resort. The kernel will otherwise continue to support existing
defined misaligned namespaces.

Unfortunately this change needs to touch several parts of the
implementation at once:

- region/available_size: expand busy extents to current align
- region/max_available_extent: expand busy extents to current align
- namespace/size: trim free space to current align

...to keep the free space accounting conforming to the dynamic align
setting.

commit 2522afb86a8cceba0f67dbf05772d21b76d79f06

Reported-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Reported-by: Jeff Moyer <jmoyer@redhat.com>
Reviewed-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Reviewed-by: Jeff Moyer <jmoyer@redhat.com>
Link: https://lore.kernel.org/r/158041478371.3889308.14542630147672668068.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>