[virtio] virtio: support VIRTIO_F_ORDER_PLATFORM
Message-id: <20200527032039.36912-5-jasowang@redhat.com>
Patchwork-id: 314435
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH V2 04/33] virtio: support VIRTIO_F_ORDER_PLATFORM
Bugzilla: 1771441
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Tiwei Bie <tiwei.bie@intel.com>

This patch introduces the support for VIRTIO_F_ORDER_PLATFORM.
If this feature is negotiated, the driver must use the barriers
suitable for hardware devices. Otherwise, the device and driver
are assumed to be implemented in software, that is they can be
assumed to run on identical CPUs in an SMP configuration. Thus
a weaker form of memory barriers is sufficient to yield better
performance.

It is recommended that an add-in card based PCI device offers
this feature for portability. The device will fail to operate
further or will operate in a slower emulation mode if this
feature is offered but not accepted.

Signed-off-by: Tiwei Bie <tiwei.bie@intel.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
(cherry picked from commit 45383fb0f42db3945ac6cc658704706cdae19528)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>