[security] ima: Store the measurement again when appraising a modsig
Message-id: <20200527180156.23049-72-didoming@redhat.com>
Patchwork-id: 314604
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH BZ1781925 v2 71/106] ima: Store the measurement again when appraising a modsig
Bugzilla: 1781925
RH-Acked-by: Bruno Meneguele <bmeneg@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1781925
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28848035

If the IMA template contains the "modsig" or "d-modsig" field, then the
modsig should be added to the measurement list when the file is appraised.

And that is what normally happens, but if a measurement rule caused a file
containing a modsig to be measured before a different rule causes it to be
appraised, the resulting measurement entry will not contain the modsig
because it is only fetched during appraisal. When the appraisal rule
triggers, it won't store a new measurement containing the modsig because
the file was already measured.

We need to detect that situation and store an additional measurement with
the modsig. This is done by adding an IMA_MEASURE action flag if we read a
modsig and the IMA template contains a modsig field.

commit e5092255bb3967bcc473dc86492dbbd5f7714023

Suggested-by: Mimi Zohar <zohar@linux.ibm.com>
Signed-off-by: Thiago Jung Bauermann <bauerman@linux.ibm.com>
Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>