[mm] mm/mmu_notifiers: add a get/put scheme for the registration
Message-id: <20200602062226.576333-7-airlied@redhat.com>
Patchwork-id: 316073
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3.0 PATCH v9 06/18] mm/mmu_notifiers: add a get/put scheme for the registration
Bugzilla: 1838335
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>

Bugzilla: 1838335
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28984280
Upstream: v5.4

Conflicts: kABI differences in two places, used reserved slots in mmu_notifier_ops
           to add alloc/free_notifier. Used reserved slots in mmu_notifier to add
           a lookaside structure (mmu_notifier_rh). Changes in mmu_notifier.c to
           hooks these structs up correctly.

commit 2c7933f53f6bff7656e3324ca1a04e478bdc57c1

Many places in the kernel have a flow where userspace will create some
object and that object will need to connect to the subsystem's
mmu_notifier subscription for the duration of its lifetime.

In this case the subsystem is usually tracking multiple mm_structs and it
is difficult to keep track of what struct mmu_notifier's have been
allocated for what mm's.

Since this has been open coded in a variety of exciting ways, provide core
functionality to do this safely.

This approach uses the struct mmu_notifier_ops * as a key to determine if
the subsystem has a notifier registered on the mm or not. If there is a
registration then the existing notifier struct is returned, otherwise the
ops->alloc_notifiers() is used to create a new per-subsystem notifier for
the mm.

The destroy side incorporates an async call_srcu based destruction which
will avoid bugs in the callers such as commit 6d7c3cde93c1 ("mm/hmm: fix
use after free with struct hmm in the mmu notifiers").

Since we are inside the mmu notifier core locking is fairly simple, the
allocation uses the same approach as for mmu_notifier_mm, the write side
of the mmap_sem makes everything deterministic and we only need to do
hlist_add_head_rcu() under the mm_take_all_locks(). The new users count
and the discoverability in the hlist is fully serialized by the
mmu_notifier_mm->lock.

Link: https://lore.kernel.org/r/20190806231548.25242-4-jgg@ziepe.ca
Co-developed-by: Christoph Hellwig <hch@infradead.org>
Signed-off-by: Christoph Hellwig <hch@infradead.org>
Reviewed-by: Ralph Campbell <rcampbell@nvidia.com>
Tested-by: Ralph Campbell <rcampbell@nvidia.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>