[fs] gfs2: Give up the iopen glock on contention
Message-id: <20200602191500.351449-7-agruenba@redhat.com>
Patchwork-id: 316272
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 fs v2 06/12] gfs2: Give up the iopen glock on contention
Bugzilla: 1582655
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When there's contention on the iopen glock, it means that the link count
of the corresponding inode has dropped to zero on a remote node which is
now trying to delete the inode.  In that case, try to evict the inode so
that the iopen glock will be released, which will allow the remote node
to do its job.

When the inode is still open locally, the inode's reference count won't
drop to zero and so we'll keep holding the inode and its iopen glock.
The remote node will time out its request to grab the iopen glock, and
when the inode is finally closed locally, we'll try to delete it
ourself.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
(cherry picked from commit dfc3be33cacca19a29091674bd8b9f5401c6cb48)
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>