[fs] gfs2: Try harder to delete inodes locally
Message-id: <20200602191500.351449-8-agruenba@redhat.com>
Patchwork-id: 316271
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 fs v2 07/12] gfs2: Try harder to delete inodes locally
Bugzilla: 1582655
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When an inode's link count drops to zero and the inode is cached on
other nodes, the current behavior of gfs2 is to immediately give up and
to rely on the other node(s) to delete the inode if there is iopen glock
contention.  This leads to resource group glock bouncing and the loss of
caching.  With the previous patches in place, we can fix that by not
giving up immediately.

When the inode is still open on other nodes, those nodes won't be able
to evict the inode and give up the iopen glock.  In that case, our lock
conversion request will time out.  The unlink system call will block for
the duration of the iopen lock conversion request.  We're also holding
the inode glock in EX mode for an extended duration, so other nodes
won't be able to make progress on the inode, either.

This is worse than what we had before, but we can prevent other nodes
from getting stuck by aborting our iopen locking request if there is
contention on the inode glock.  This will the the subject of a future
patch.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
(cherry picked from commit 8c54791b55bcbdf15c3a4f3135b02625ee6979c5)
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>