[fs] gfs2: Smarter iopen glock waiting
Message-id: <20200602191500.351449-13-agruenba@redhat.com>
Patchwork-id: 316277
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 fs v2 12/12] gfs2: Smarter iopen glock waiting
Bugzilla: 1582655
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When trying to upgrade the iopen glock from a shared to an exclusive lock in
gfs2_evict_inode, abort the wait if there is contention on the corresponding
inode glock: in that case, the inode must still be in active use on another
node, and we're not guaranteed to get the iopen glock anytime soon.

To make this work even better, when we notice contention on the iopen glock and
we can't evict the corresponsing inode and release the iopen glock immediately,
poke the inode glock.  The other node(s) trying to acquire the lock can then
abort instead of timing out.

Thanks to Heinz Mauelshagen for pointing out a locking bug in a previous
version of this patch.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
(cherry picked from commit cc684583b2dd3a5818e1951cbc8caa699a3055a0)
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>