[x86] kvm: nvmx: Add helper to handle TLB flushes on nested VM-Enter/VM-Exit
Message-id: <20200604183347.508152-49-vkuznets@redhat.com>
Patchwork-id: 316839
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH 048/235] KVM: nVMX: Add helper to handle TLB flushes on nested VM-Enter/VM-Exit
Bugzilla: 1841490
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Sean Christopherson <sean.j.christopherson@intel.com>

Add a helper to determine whether or not a full TLB flush needs to be
performed on nested VM-Enter/VM-Exit, as the logic is identical for both
flows and needs a fairly beefy comment to boot.  This also provides a
common point to make future adjustments to the logic.

Handle vpid12 changes the new helper as well even though it is specific
to VM-Enter.  The vpid12 logic is an extension of the flushing logic,
and it's worth the extra bool parameter to provide a single location for
the flushing logic.

Cc: Liran Alon <liran.alon@oracle.com>
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Message-Id: <20200320212833.3507-24-sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 50b265a4eee0d9fb9581f1a300742515b80dffda)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>