[x86] kvm: nsvm: Report interrupts as allowed when in L2 and exit-on-interrupt is set
Message-id: <20200604183347.508152-133-vkuznets@redhat.com>
Patchwork-id: 316897
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH 132/235] KVM: nSVM: Report interrupts as allowed when in L2 and exit-on-interrupt is set
Bugzilla: 1841490
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

Report interrupts as allowed when the vCPU is in L2 and L2 is being run with
exit-on-interrupts enabled and EFLAGS.IF=1 (either on the host or on the guest
according to VINTR).  Interrupts are always unblocked from L1's perspective
in this case.

While moving nested_exit_on_intr to svm.h, use INTERCEPT_INTR properly instead
of assuming it's zero (which it is of course).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit fc6f7c03ad80693c953d5cdacfad41f174289531)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>