[mm] mm: revert remove MEMORY_DEVICE_PUBLIC support in swap.c
Message-id: <20200604201526.20430-3-ddutile@redhat.com>
Patchwork-id: 317169
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL-8.3 2/5] mm: revert remove MEMORY_DEVICE_PUBLIC support in swap.c
Bugzilla: 1816441
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Christoph von Recklinghausen <crecklin@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1816441

Upstream Status: RHEL-8 only

Revert hunk of upstream commit 25b2995a35b609119cf96
for swap.c that was backported to RHEL-8.2.
Upstream part of this commit not in git log history of mm/swap.c
Apparently this hunk of commit 25b2995a35b609119cf96 was not removed
(in a merge with the next patch upstream?)
so that the next commit in this series would be applied, effectively
removing DEVICE_PUBLIC support (intent of original commit), so it could
be replaced with a more generic device-zone check.

Another approach would be to mangle the next backport'd commit, but this
patch and the next better 'follows' the reason for the change sequence.

*NOTE*: This revert, as-is, will break a kernel bisect build
	because is_device_public_page() is not added back to
	include/linux/mm.h.  To avoid adding a rhel-8-only hunk
	to the next patch to remove it, since that patch will remove
	the only use of is_device_public_page(), just note bisect-build
	error so a developer can skip to next bisect.

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>