[kernel] perf/core: Use min_heap in visit_groups_merge()
Message-id: <20200605130702.31420-11-mpetlan@redhat.com>
Patchwork-id: 317390
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 010/112] perf/core: Use min_heap in visit_groups_merge()
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit 6eef8a7116deae0706ba6d897c0d7dd887cd2be2
Author: Ian Rogers <irogers@google.com>
Date: Thu Feb 13 23:51:30 2020 -0800

description
===========
visit_groups_merge will pick the next event based on when it was
inserted in to the context (perf_event group_index). Events may be per CPU
or for any CPU, but in the future we'd also like to have per cgroup events
to avoid searching all events for the events to schedule for a cgroup.
Introduce a min heap for the events that maintains a property that the
earliest inserted event is always at the 0th element. Initialize the heap
with per-CPU and any-CPU events for the context.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>