[kernel] perf/cgroup: Order events in RB tree by cgroup id
Message-id: <20200605130702.31420-14-mpetlan@redhat.com>
Patchwork-id: 317393
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 013/112] perf/cgroup: Order events in RB tree by cgroup id
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit 95ed6c707f26a727a29972b60469630ae10d579c
Author: Ian Rogers <irogers@google.com>
Date: Thu Feb 13 23:51:33 2020 -0800

description
===========
If one is monitoring 6 events on 20 cgroups the per-CPU RB tree will
hold 120 events. The scheduling in of the events currently iterates
over all events looking to see which events match the task's cgroup or
its cgroup hierarchy. If a task is in 1 cgroup with 6 events, then 114
events are considered unnecessarily.

This change orders events in the RB tree by cgroup id if it is present.
This means scheduling in may go directly to events associated with the
task's cgroup if one is present. The per-CPU iterator storage in
visit_groups_merge is sized sufficent for an iterator per cgroup depth,
where different iterators are needed for the task's cgroup and parent
cgroups. By considering the set of iterators when visiting, the lowest
group_index event may be selected and the insertion order group_index
property is maintained. This also allows event rotation to function
correctly, as although events are grouped into a cgroup, rotation always
selects the lowest group_index event to rotate (delete/insert into the
tree) and the min heap of iterators make it so that the group_index order
is maintained.

Conflicts:
==========
Since we haven't backported 67c0496e87d193 ("kernfs: convert kernfs_node->id
from union kernfs_node_id to u64"), we have to "cast" the union kernfs_node->id
to u64.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>