[tools] perf report/top TUI: Support hotkey 'a' for annotation of unresolved addresses
Message-id: <20200605130702.31420-70-mpetlan@redhat.com>
Patchwork-id: 317451
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 069/112] perf report/top TUI: Support hotkey 'a' for annotation of unresolved addresses
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit ec0479a63b7657e57665a9fc81d11a99131cfc62
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Thu Feb 27 12:39:39 2020 +0800

description
===========
In previous patch, we have supported the annotation functionality even
without symbols.

For this patch, it supports the hotkey 'a' on address in report view.
Note that, for branch mode, we only support the annotation for "branch
to" address.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>