[tools] perf tools: Add file-handle feature test
Message-id: <20200605130702.31420-91-mpetlan@redhat.com>
Patchwork-id: 317472
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 090/112] perf tools: Add file-handle feature test
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit 49f550ea87c73671ee4d5e08820e0976894dd610
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Apr 2 10:52:49 2020 +0900

description
===========
The file handle (FHANDLE) support is configurable so some systems might not
have it.  So add a config feature item to check it on build time so that we
don't add the cgroup tracking feature based on that.

Committer notes:

Had to make the test use the same construct as its later use in
synthetic-events.c, in the next patch in this series. i.e. make it be:

	struct {
		struct file_handle fh;
		uint64_t cgroup_id;
	} handle;

To cope with:

    CC       /tmp/build/perf/util/cloexec.o
  util/synthetic-events.c:428:22: error: field 'fh' with   CC       /tmp/build/perf/util/call-path.o
  variable sized type 'struct file_handle' not at the end of a struct or class is a GNU
        extension [-Werror,-Wgnu-variable-sized-type-not-at-end]
                  struct file_handle fh;
                                     ^
  1 error generated.

Deal with this at some point, i.e. investigate if the right thing is to
remove that -Wgnu-variable-sized-type-not-at-end from our CFLAGS, for
now do the test the same way as it is used looks more sensible.

[ split from a larger patch, removed blank line at EOF ]

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>