[tools] perf build-test: Honour JOBS to override detection of number of cores
Message-id: <20200605130702.31420-99-mpetlan@redhat.com>
Patchwork-id: 317480
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 098/112] perf build-test: Honour JOBS to override detection of number of cores
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit 7b1642f2fc1e1f20948e806b7ce319f660633342
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Mar 30 09:32:41 2020 -0300

description
===========
When one does:

  $ make -C tools/perf build-test

The makefile in tools/perf/tests/ will, just like the main one, detect
how many cores are in the system and use it with -j.

Sometimes we may need to override that, for instance, when using
icecream or distcc to use multiple machines in the build process, then
we need to, as with the main makefile, use:

  $ make JOBS=N -C tools/perf build-test

Fix the tests makefile to honour that.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>