[tools] perf tools: Support Python 3.8+ in Makefile
Message-id: <20200605130702.31420-109-mpetlan@redhat.com>
Patchwork-id: 317490
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 108/112] perf tools: Support Python 3.8+ in Makefile
Bugzilla: 1841130
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1841130
https://bugzilla.redhat.com/show_bug.cgi?id=1841130

upstream
========
commit b9c9ce4e598e012ca7c1813fae2f4d02395807de
Author: Sam Lunt <samueljlunt@gmail.com>
Date: Fri Jan 31 12:11:23 2020 -0600

description
===========
Python 3.8 changed the output of 'python-config --ldflags' to no longer
include the '-lpythonX.Y' flag (this apparently fixed an issue loading
modules with a statically linked Python executable).  The libpython
feature check in linux/build/feature fails if the Python library is not
included in FEATURE_CHECK_LDFLAGS-libpython variable.

This adds a check in the Makefile to determine if PYTHON_CONFIG accepts
the '--embed' flag and passes that flag alongside '--ldflags' if so.

tools/perf is the only place the libpython feature check is used.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>