RHMAINTAINERS: Remove Partner Engineers from RHMAINTAINERS
We are in discussions with Partner Management to see if we can ask Partner
Engineers to include themselves in the ARK RHMAINTAINERS.  For now,
to make the transition of Red Hat Employees in RHMAINTAINERS to ARK, move
the Partner entries into their own file.  This new file, included in the
MAINTAINERS directory will allow for the current get_maintainer.conf file
to continue working for RHEL8.

Remove the Partner Engineers from RHMAINTAINERS and move them into a
separate PARTNERS file.

RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: John W. Linville <linville@redhat.com>

Signed-off-by: Prarit Bhargava <prarit@redhat.com>