[netdrv] ice: Don't reset and rebuild for Tx timeout on PFC enabled queue
Message-id: <80b29461e2fd00cd7bffe3948d186eacb0c7c5a8.1591891905.git.jtoppins@redhat.com>
Patchwork-id: 320029
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL-8.3 04/17] ice: Don't reset and rebuild for Tx timeout on PFC enabled queue
Bugzilla: 1781405
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

When there's a Tx timeout for a queue which belongs to a PFC enabled TC,
then it's not because the queue is hung but because PFC is in action.

In PFC, peer sends a pause frame for a specified period of time when its
buffer threshold is exceeded (due to congestion). Netdev on the other
hand checks if ACK is received within a specified time for a TX packet, if
not, it'll invoke the tx_timeout routine.

Signed-off-by: Avinash JD <avinash.dayanand@intel.com>
Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 610ed0e93e7dd107afc2398d54de6b6fe9513f8d)
Bugzilla: 1781405
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=29271836
Tested: build only, provided to partner for testing
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>