[s390] cio: Update SCSW if it points to the end of the chain
Message-id: <20200618145311.270918-23-cohuck@redhat.com>
Patchwork-id: 320519
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt 22/74] s390/cio: Update SCSW if it points to the end of the chain
Bugzilla: 1660917
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Eric Farman <farman@linux.ibm.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1660917
Upstream Status: upstream
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=29479598

Per the POPs [1], when processing an interrupt the SCSW.CPA field of an
IRB generally points to 8 bytes after the last CCW that was executed
(there are exceptions, but this is the most common behavior).

In the case of an error, this points us to the first un-executed CCW
in the chain.  But in the case of normal I/O, the address points beyond
the end of the chain.  While the guest generally only cares about this
when possibly restarting a channel program after error recovery, we
should convert the address even in the good scenario so that we provide
a consistent, valid, response upon I/O completion.

[1] Figure 16-6 in SA22-7832-11.  The footnotes in that table also state
that this is true even if the resulting address is invalid or protected,
but moving to the end of the guest chain should not be a surprise.

Signed-off-by: Eric Farman <farman@linux.ibm.com>
Message-Id: <20190514234248.36203-2-farman@linux.ibm.com>
Reviewed-by: Farhan Ali <alifm@linux.ibm.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>
(cherry picked from commit 15f0eb3d6a62b22b27449329f04ee9aa0d218c6a)
Signed-off-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>