[fs] ceph: flush release queue when handling caps for unknown inode
Message-id: <20200623110929.7600-3-jlayton@redhat.com>
Patchwork-id: 321397
Patchwork-instance: patchwork
O-Subject: [BZ#1847528 RHEL8.3 PATCH 02/23] ceph: flush release queue when handling caps for unknown inode
Bugzilla: 1847528
RH-Acked-by: Zheng Yan <zyan@redhat.com>
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Ilya Dryomov <idryomov@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

It's possible for the VFS to completely forget about an inode, but for
it to still be sitting on the cap release queue. If the MDS sends the
client a cap message for such an inode, it just ignores it today, which
can lead to a stall of up to 5s until the cap release queue is flushed.

If we get a cap message for an inode that can't be located, then go
ahead and flush the cap release queue.

Cc: stable@vger.kernel.org
URL: https://tracker.ceph.com/issues/45532
Fixes: 1e9c2eb6811e ("ceph: delete stale dentry when last reference is dropped")
Reported-and-Tested-by: Andrej Filipčič <andrej.filipcic@ijs.si>
Suggested-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit fb33c114d3ed5bdac230716f5b0a93b56b92a90d)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>