[fs] ceph: convert mdsc->cap_dirty to a per-session list
Message-id: <20200623110929.7600-13-jlayton@redhat.com>
Patchwork-id: 321409
Patchwork-instance: patchwork
O-Subject: [BZ#1847528 RHEL8.3 PATCH 12/23] ceph: convert mdsc->cap_dirty to a per-session list
Bugzilla: 1847528
RH-Acked-by: Zheng Yan <zyan@redhat.com>
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Ilya Dryomov <idryomov@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

This is a per-sb list now, but that makes it difficult to tell when
the cap is the last dirty one associated with the session. Switch
this to be a per-session list, but continue using the
mdsc->cap_dirty_lock to protect the lists.

This list is only ever walked in ceph_flush_dirty_caps, so change that
to walk the sessions array and then flush the caps for inodes on each
session's list.

If the auth cap ever changes while the inode has dirty caps, then
move the inode to the appropriate session for the new auth_cap. Also,
ensure that we never remove an auth cap while the inode is still on the
s_cap_dirty list.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 1cf03a68e791b1673bc4daaa88a0820f34f538f8)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>